requirements.txt

streamlit
easyocr
pillow
pandas
xlsxwriter
numpy
